fix(用例管理): 修复脑图导入相关以及脑图编辑相关问题
--bug=1045261 --user=郭雨琦 https://www.tapd.cn/55049933/bugtrace/bugs/view/1155049933001045261
--bug=1045260 --user=郭雨琦 https://www.tapd.cn/55049933/bugtrace/bugs/view/1155049933001045260
--bug=1045259 --user=郭雨琦 https://www.tapd.cn/55049933/bugtrace/bugs/view/1155049933001045259
--bug=1045258 --user=郭雨琦 https://www.tapd.cn/55049933/bugtrace/bugs/view/1155049933001045258
--bug=1045257 --user=郭雨琦 https://www.tapd.cn/55049933/bugtrace/bugs/view/1155049933001045257
--bug=1045233 --user=郭雨琦 https://www.tapd.cn/55049933/bugtrace/bugs/view/1155049933001045233
--bug=1045219 --user=郭雨琦 https://www.tapd.cn/55049933/bugtrace/bugs/view/1155049933001045219
--bug=1045216 --user=郭雨琦 https://www.tapd.cn/55049933/bugtrace/bugs/view/1155049933001045216
--bug=1045214 --user=郭雨琦 https://www.tapd.cn/55049933/bugtrace/bugs/view/1155049933001045214
--bug=1045209 --user=郭雨琦 https://www.tapd.cn/55049933/bugtrace/bugs/view/1155049933001045209
--bug=1045203 --user=郭雨琦 https://www.tapd.cn/55049933/bugtrace/bugs/view/1155049933001045203

+++ backend/services/case-management/src/main/java/io/metersphere/functional/service/FunctionalCaseMinderService.java
@@ -1184,7 +1184,13 @@ public class FunctionalCaseMinderService {
             }
             List<MinderOptionDTO> caseModuleOptionDTOS = resourceMap.get(Translator.get("minder_extra_node.module"));
             if (CollectionUtils.isNotEmpty(caseModuleOptionDTOS)) {
-                List<String> moduleIds = caseModuleOptionDTOS.stream().map(MinderOptionDTO::getId).toList();
+                List<String> moduleIds = new ArrayList<>(caseModuleOptionDTOS.stream().map(MinderOptionDTO::getId).toList());
+                if (moduleIds.contains("NONE")) {
+                    throw new MSException(Translator.get("all.module.default.name.cut_error"));
+                }
+                if (moduleIds.contains("root")) {
+                    throw new MSException(Translator.get("functional_case.module.default.name.cut_error"));
+                }
                 List<FunctionalCase> functionalCases = functionalCaseModuleService.deleteModuleByIds(moduleIds, new ArrayList<>(), userId);
                 functionalCaseModuleService.batchDelLog(functionalCases, request.getProjectId());
                 List<String> finalCaseIds = caseIds;
